Change not throw exception even if the `yamorySbtScriptUrl` or `yamoryYarnScriptUrl` is not set.

+++ src/main/scala/io/github/kijuky/sbt/plugins/yamory/YamoryPlugin.scala
@@ -63,29 +63,27 @@ object YamoryPlugin extends AutoPlugin {
       yamoryApiKey.nonEmpty,
       "YAMORY_API_KEY is empty. set 'yamoryApiKey' setting."
     )
-    require(
-      yamorySbtScriptUrl.nonEmpty,
-      "yamory sbt script url is empty. set 'yamorySbtScriptUrl' setting."
-    )
 
-    val dependenciesFile = Files.createTempFile("sbt", ".txt").toFile
-    val yamorySbtScriptFile = Files.createTempFile("sbt", ".sh").toFile
-    Seq(dependenciesFile, yamorySbtScriptFile).foreach(_.deleteOnExit())
-    try {
-      val dependencies = (Compile / dependencyTree / asString).value
-      val dependenciesLog =
-        dependencies.split("\n").map("[info] ".+).mkString("\n")
-      IO.write(dependenciesFile, dependenciesLog, IO.utf8)
-      url(yamorySbtScriptUrl) #> yamorySbtScriptFile !
-      val yamorySbtScriptFilePath = yamorySbtScriptFile.getAbsolutePath
-      s"chmod +x $yamorySbtScriptFilePath" #&& (dependenciesFile #> Process(
-        Seq("bash", "-c", yamorySbtScriptFilePath),
-        None,
-        "PROJECT_GROUP_KEY" -> projectGroupKey,
-        "YAMORY_API_KEY" -> yamoryApiKey
-      )) !
-    } finally {
-      Seq(dependenciesFile, yamorySbtScriptFile).foreach(_.delete())
+    if (yamorySbtScriptUrl.nonEmpty) {
+      val dependenciesFile = Files.createTempFile("sbt", ".txt").toFile
+      val yamorySbtScriptFile = Files.createTempFile("sbt", ".sh").toFile
+      Seq(dependenciesFile, yamorySbtScriptFile).foreach(_.deleteOnExit())
+      try {
+        val dependencies = (Compile / dependencyTree / asString).value
+        val dependenciesLog =
+          dependencies.split("\n").map("[info] ".+).mkString("\n")
+        IO.write(dependenciesFile, dependenciesLog, IO.utf8)
+        url(yamorySbtScriptUrl) #> yamorySbtScriptFile !
+        val yamorySbtScriptFilePath = yamorySbtScriptFile.getAbsolutePath
+        s"chmod +x $yamorySbtScriptFilePath" #&& (dependenciesFile #> Process(
+          Seq("bash", "-c", yamorySbtScriptFilePath),
+          None,
+          "PROJECT_GROUP_KEY" -> projectGroupKey,
+          "YAMORY_API_KEY" -> yamoryApiKey
+        )) !
+      } finally {
+        Seq(dependenciesFile, yamorySbtScriptFile).foreach(_.delete())
+      }
     }
   }
 
@@ -104,34 +102,32 @@ object YamoryPlugin extends AutoPlugin {
       "YAMORY_API_KEY is empty. set 'yamoryApiKey' setting."
     )
     require(
-      yamoryYarnScriptUrl.nonEmpty,
-      "yamory yarn script url is empty. set 'yamoryYarnScriptUrl' setting."
-    )
-    require(
       yamoryYarnManifest.nonEmpty,
       "yamory yarn manifest is empty. set 'yamoryYarnManifest' setting."
     )
 
-    val yamoryYarnScriptFile = Files.createTempFile("sbt", ".sh").toFile
-    yamoryYarnScriptFile.deleteOnExit()
-    try {
-      url(yamoryYarnScriptUrl) #> yamoryYarnScriptFile !
-      val yamoryYarnScriptFilePath = yamoryYarnScriptFile.getAbsolutePath
-      s"chmod +x $yamoryYarnScriptFilePath" #&& Process(
-        Seq(
-          "bash",
-          "-c",
-          yamoryYarnScriptFilePath,
-          "--",
-          "--manifest",
-          yamoryYarnManifest
-        ),
-        None,
-        "PROJECT_GROUP_KEY" -> projectGroupKey,
-        "YAMORY_API_KEY" -> yamoryApiKey
-      ) !
-    } finally {
-      yamoryYarnScriptFile.delete()
+    if (yamoryYarnScriptUrl.nonEmpty) {
+      val yamoryYarnScriptFile = Files.createTempFile("sbt", ".sh").toFile
+      yamoryYarnScriptFile.deleteOnExit()
+      try {
+        url(yamoryYarnScriptUrl) #> yamoryYarnScriptFile !
+        val yamoryYarnScriptFilePath = yamoryYarnScriptFile.getAbsolutePath
+        s"chmod +x $yamoryYarnScriptFilePath" #&& Process(
+          Seq(
+            "bash",
+            "-c",
+            yamoryYarnScriptFilePath,
+            "--",
+            "--manifest",
+            yamoryYarnManifest
+          ),
+          None,
+          "PROJECT_GROUP_KEY" -> projectGroupKey,
+          "YAMORY_API_KEY" -> yamoryApiKey
+        ) !
+      } finally {
+        yamoryYarnScriptFile.delete()
+      }
     }
   }
 }
